Move onboarding blocs
