+++ lib/view/onboarding/auth_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:mobile/flavors.dart';
 import 'package:mobile/view/components/flavor_title_badge.dart';
-import 'package:mobile/view/onboarding/auth_navigation_bloc.dart';
+import 'package:mobile/bloc/onborading/auth_navigation_bloc.dart';
 import 'package:mobile/view/onboarding/login_screen.dart';
 import 'package:mobile/view/onboarding/signup_screen.dart';
