+++ test/view/onboarding/auth_navigation_bloc_test.dart
@@ -1,6 +1,6 @@
 import 'package:bloc_test/bloc_test.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mobile/view/onboarding/auth_navigation_bloc.dart';
+import 'package:mobile/bloc/onborading/auth_navigation_bloc.dart';
 
 void main() {
   group('AuthNavigationBloc test', () {
